+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod collector;
 pub mod commands;
 pub mod config;
+pub mod daemon;
 pub mod keyword;
 pub mod params;
 pub mod runner;